Refactor: header에 로그인 모달 연결 & 마이페이지 리다이렉트 수정
