+++ src/app/(auth)/sign-in/page.tsx
@@ -1,11 +1,11 @@
 'use client';
 
 // 테스트 페이지 입니다.
+import { Button, Modal } from '@/components';
 import classNames from 'classnames/bind';
 import { useState } from 'react';
-import { Modal, Button } from '@/components';
+import SignInModal from '../../../components/SignInModal/SignInModal';
 import styles from './page.module.scss';
-import SignInModal from './_component/SignInModal';
 
 const cn = classNames.bind(styles);
