+++ src/app/my-info/(account)/addresses/layout.tsx
@@ -1,8 +1,11 @@
+import { HydrationBoundary, QueryClient, dehydrate } from '@tanstack/react-query';
 import classNames from 'classnames/bind';
+import { redirect } from 'next/navigation';
 import { ReactNode } from 'react';
 
 import { getAddresses } from '@/api/shippingAPI';
-import { HydrationBoundary, QueryClient, dehydrate } from '@tanstack/react-query';
+import { ROUTER } from '@/constants/route';
+import { getCookie } from '@/libs/manageCookie';
 import { AddressesHeader } from './_components';
 
 import styles from './layout.module.scss';
@@ -14,6 +17,12 @@ interface AddressesLayoutProps {
 }
 
 export default async function AddressesLayout({ children }: AddressesLayoutProps) {
+  const token = await getCookie('accessToken');
+
+  if (!token) {
+    redirect(ROUTER.MAIN);
+  }
+
   const queryClient = new QueryClient();
 
   await queryClient.prefetchQuery({ queryKey: ['addressesData'], queryFn: getAddresses });
