+++ src/app/my-info/layout.tsx
@@ -1,11 +1,6 @@
-import { ROUTER } from '@/constants/route';
 import classNames from 'classnames/bind';
 import { ReactNode } from 'react';
-
-import { getCookie } from '@/libs/manageCookie';
-import { redirect } from 'next/navigation';
 import { SNB } from './_components';
-
 import styles from './layout.module.scss';
 
 const cn = classNames.bind(styles);
@@ -14,13 +9,7 @@ interface MyInfoLayoutProps {
   children: ReactNode;
 }
 
-export default async function MyInfoLayout({ children }: MyInfoLayoutProps) {
-  const token = await getCookie('accessToken');
-
-  if (!token) {
-    redirect(ROUTER.MAIN);
-  }
-
+export default function MyInfoLayout({ children }: MyInfoLayoutProps) {
   return (
     <section className={cn('layout')}>
       <SNB />
