+++ src/components/Header/HeaderParts/LoginButton.tsx
@@ -1,17 +1,18 @@
 'use client';
 
 import classNames from 'classnames/bind';
+
 import styles from './AuthButton.module.scss';
 
 const cn = classNames.bind(styles);
 
-export default function LoginButton() {
-  const handleClickButton = () => {
-    /* Login modal */
-  };
+interface LoginButtonProps {
+  onClick: () => void;
+}
 
+export default function LoginButton({ onClick }: LoginButtonProps) {
   return (
-    <button className={cn('button')} type='button' onClick={handleClickButton}>
+    <button className={cn('button')} type='button' onClick={onClick}>
       로그인
     </button>
   );
