+++ src/components/SignInModal/SignInModal.tsx
@@ -1,15 +1,16 @@
 import classNames from 'classnames/bind';
-import { FieldValues, SubmitHandler, useForm } from 'react-hook-form';
 import Link from 'next/link';
+import { FieldValues, SubmitHandler, useForm } from 'react-hook-form';
 import { toast } from 'react-toastify';
 
-import { InputField, Button } from '@/components';
-import { GitHubIcon, GoogleIcon, KakaoIcon } from '@/public/index';
 import { postSignin } from '@/api/authAPI';
-import { setCookie } from '@/libs/manageCookie';
-import type { FetchSignInInfoTypes } from '@/types/authTypes';
 import { ROUTER } from '@/constants/route';
+import { setCookie } from '@/libs/manageCookie';
+import { GitHubIcon, GoogleIcon, KakaoIcon } from '@/public/index';
+import type { FetchSignInInfoTypes } from '@/types/authTypes';
 
+import Button from '../Buttons/Button/Button';
+import InputField from '../InputField/InputField';
 import styles from './SigninModal.module.scss';
 
 const cn = classNames.bind(styles);
