+++ src/components/parts/Input.module.scss
@@ -11,19 +11,19 @@
 
 @mixin sm {
   height: 4rem;
-  padding: 0.8rem 3.2rem 0.8rem 1.6rem;
+  padding: 0.8rem 4rem 0.8rem 1.6rem;
   @include pretendard-16-400;
 }
 
 @mixin md {
   height: 4.4rem;
-  padding: 1rem 3.2rem 1rem 1.6rem;
+  padding: 1rem 4rem 1rem 1.6rem;
   @include pretendard-16-400;
 }
 
 @mixin lg {
   height: 4.4rem;
-  padding: 1rem 3.2rem 1rem 1.6rem;
+  padding: 1rem 4rem 1rem 1.6rem;
   @include pretendard-18-400;
 }
